More Mongoose connections - updated wavesurfer look
