+++ backend/controllers/calls.js
@@ -1,6 +1,6 @@
 
 const { ObjectId } = require('mongodb');
-var { callModel: Call } = require("../models/call");
+const Call = require("../models/call");
 const Group = require("../models/group");
 const opentelemetry = require('@opentelemetry/api');
 var defaultNumResults = 50;
