+++ backend/models/frozenCall.js
@@ -1,0 +1,4 @@
+var mongoose = require('mongoose');
+const frozenCallSchema = require('./frozenCallSchema');
+
+module.exports = mongoose.model('FrozenCall', frozenCallSchema);
