+++ backend/models/talkgroup.js
@@ -1,16 +1,4 @@
 var mongoose = require('mongoose');
-
-var talkgroupSchema = mongoose.Schema({
-  shortName: String,
-  userId: String,
-  num: Number,
-  mode: String,
-  alpha: String,
-  description: String,
-  tag: String,
-  group: String,
-  priority: Number
-});
-
+const talkgroupSchema = require('./talkgroupSchema');
 
 module.exports = mongoose.model('Talkgroup', talkgroupSchema);
