+++ frontend/src/Call/components/MediaPlayer.css
@@ -88,3 +88,7 @@ background-color: #5c5c5c;
         color: white;
       }
 }
+
+i.icon {
+    font-size: inherit;
+}
